Allow Rails Middleware Management by User
WOVN.rb railtie changes so that `Wovnrb::Interceptor` middleware is added
only if user did not add it themselves. This change allows for
configuring middlewares that places before `Wovnrb::Interceptor` to decide
whether or not `wovn_disable` should be set to true.
`previous_middleware` setting is no longer needed.

Following is a Rails setup example.
  `config/application.rb`
    ```

    class TestMiddleware
      def initialize(app)
        @app = app
      end

      def call(env)
        Rack::Request.new(env).update_param('wovn_disable', true)
        return @app.call(env)
      end
    end

    module MyApp
      class Application < Rails::Application
        # ...
        config.middleware.insert_before(0, TestMiddleware)
        config.middleware.insert_after(TestMiddleware, Wovnrb::Interceptor)
        # ...
      end
    end
    ```