[IS-376] Add support for `lang_param_name` (#149)
* Add `lang_param_name` option

`lang_param_name` option allows user to customize the query parameter
user to store the language information when `url_pattern_name` is set to
`query`. Previous, `wovn` was the only available value for the name of
the query parameter to store the language information. `wovn` remains
the default value for retro compatibility.